Fixed a nasty bug in 64-bit integer handling.
I had forgotten to set the precision when constructing an HDF5 type from and I3Datatype. The HDF5 library reduces precision when an integer is truncated (e.g. int -> short), but does not expand it automatically when the width of the type is increased. For 64-bit integers, this created an 8-byte integer with 4-byte precision. When reading the table back out, the HDF5 library noted that there were only 4 significant bytes and only wrote out those four bytes. Pytables assumes that all integers have precision 8*width, so does not zero the memory it hands to the HDF5 library to be filled with data. This resulted in the upper 32 bits being filled with random gibberish that changed each time the data were read out. Phew.

git-svn-id: http://code.icecube.wisc.edu/svn/projects/hdfwriter/trunk@62683 16731396-06f5-0310-8873-f7f720988828