Implemented buffered reads for I3HDFTable.
Reads are consolidated into the 8 times the chunk size (e.g. 1 MB). Any time a read request falls outside the bounds of the read cache, the cache is invalidated and a new block of chunks is read from disk. This optimization targets sequential reads, as happen during transcription or table merging. Overall, the table transcriber should now be usable for merging, and should at least not be significantly slower than the pytables-based merging script.

git-svn-id: http://code.icecube.wisc.edu/svn/projects/hdfwriter/trunk@62811 16731396-06f5-0310-8873-f7f720988828